requirements.txt
google-adk==1.14.0
yfinance==0.2.65


